+++ CHANGELOG.rst
@@ -21,6 +21,7 @@ Removed
 
 Fixed
 =====
+- [v0x01] Fixed method to unpack error messages.
 
 Security
 ========